refactor and make highdist, qaa and run left
